requirements.txt
streamlit
pdfplumber
python-dotenv
fuzzywuzzy
requests
huggingface_hub
huggingface
langchain